+++ .idea/.gitignore
@@ -2,3 +2,4 @@
 /shelf/
 /workspace.xml
 /misc.xml
+misc.xml
