+++ src/modules/Tasks/Timed/TimedTask.java
@@ -4,4 +4,7 @@ import modules.Tasks.Task;
 
 public class TimedTask extends Task {
 
+    public TimedTask(){
+        super();
+    }
 }
